### 2019-01-11 #1165 @whedon
[PRE REVIEW]: HODLRlib: A Library for Hierarchical Matrices
**Submitting author:** @sivaramambikasaran (<a href="http://orcid.org/0000-0003-2978-6281">Sivaram Ambikasaran</a>)
**Repository:** <a href="https://github.com/sivaramambikasaran/HODLR" target ="_blank">https://github.com/sivaramambikasaran/HODLR</a>
**Version:** 3.1415
