### 2019-01-12 #1165 @labarba
Perhaps for this submission, I could edit and @poulson could review? He can also observe the editing process to learn the ropes.
