### 2019-01-12 #1167 @whedon
[REVIEW]: HODLRlib: A Library for Hierarchical Matrices
**Submitting author:** @sivaramambikasaran (<a href="http://orcid.org/0000-0003-2978-6281">Sivaram Ambikasaran</a>)
**Repository:** <a href="https://github.com/sivaramambikasaran/HODLR" target ="_blank">https://github.com/sivaramambikasaran/HODLR</a>
**Version:** v3.1415.1
