### 2019-01-12 #1167 @jedbrown
@poulson :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the HODLRlib repository). I'll be watching this thread if you have any questions.
