### 2019-01-13 sivaramambikasaran/HODLR#22 @poulson
install.sh needs to download a newer version of CMake for `add_compile_definitions` support
`install.sh` is currently downloading CMake 3.3.2, but `add_compile_definitions` (which is currently used by HODLR) requires CMake >= 3.12. As is, attempting to configure for the unit tests via
```
../deps/cmake/bin/cmake .. -DCMAKE_BUILD_TYPE=COVERAGE -DINPUT_FILE=../test/test_HODLR.cpp
```
yields an error
```
CMake Error at CMakeLists.txt:24 (add_compile_definitions):
  Unknown CMake command "add_compile_definitions".
```

I therefore recommend using the following replacement within `install.sh` (which I have verified to fix this issue):
```
    CMAKE_URL="https://cmake.org/files/v3.12/cmake-3.12.4-Linux-x86_64.tar.gz"
```

(As an aside, I recommend making `install.sh` executable so that `./install.sh` is possible and adding an `echo` warning the user that the dependencies may take some time to download and build.)
