### 2019-01-13 #1167 @poulson
I have filed https://github.com/sivaramambikasaran/HODLR/issues/22 to suggest `install.sh` downloading a more recent version of CMake since the current choice does not support all of the features needed by the project (i.e., `add_compile_definitions`).
