### 2019-01-13 a76ebd0
Edit to CMakeLists.txt to fix issue in testing
