### 2019-01-14 b052b50
Update installation.rst
