### 2019-01-14 c89013f
Update install.sh
