### 2019-01-14 sivaramambikasaran/HODLR#22 @sivaramambikasaran
@poulson These have now been addressed. Also, a typo in the code in the docs has been updated.
