### 2019-01-14 #1167 @sivaramambikasaran
@poulson These have now been addressed.
