### 2019-01-14 31b8c4f
Update install.sh
