### 2019-01-14 fa3d7b6
Change to make the plotTree function to take the output filename and extension as an argument
