+++ examples/tutorial.cpp
@@ -132,7 +132,7 @@ int main(int argc, char* argv[])
 
     // These are mainly used in development and debugging:
     // Used to visualize the rank structure of the considered kernel:
-    // T->plotTree();
+    // T->plotTree("plot.svg");
     // Prints the details of all the nodes in the tree:
     // T->printTreeDetails();
 
