### 2019-01-19 #1167 @poulson
@sivaramambikasaran Thank you for fixing that issue; I have verified that a from-scratch usage of `install.sh` now builds the library! (Though I had to manually install `dvipng`.)

I am now noticing that running a standard build and then executing `./tutorial` leads to a segfault. It seems that one must manually specify the matrix height (`N`), leaf cutoff size (`M`), off-diagonal block rank (`dim`), and -log10(tolerance) as the first 4 arguments to `tutorial` to avoid a segfault. I would recommend using default versions of these parameters (or exiting with an error message) rather than segfaulting (and perhaps making note in the installation notes next to the cmake instructions).

I am also wondering what the recommended example is, as the runtimes for HODLR seem to be significantly slower than dense linear algebra up to a matrix dimension of 4000. For example:
```
~/Source/JOSS/HODLR/build-full$ ./tutorial 4000 128 10 12
========================= Problem Parameters =========================
Matrix Size                        :4000
Leaf Size                          :128
Number of Levels in Tree           :4
Dimensionality                     :10
Tolerance                          :1e-12

========================= Assembly Time =========================
Time for assembly in HODLR form    :37.7604
Time for direct matrix generation  :0.513778
Magnitude of Speed-Up              :0.0136063

========================= Matrix-Vector Multiplication =========================
Time for MatVec in HODLR form      :0.0915819
Time for direct MatVec             :0.0725777
Magnitude of Speed-Up              :0.79249
Error in the solution is           :3.47459e-16

========================= Factorization =========================
Time to factorize HODLR form       :39.0081
Time for LU Factorization          :6.82005
Magnitude of Speed-Up              :0.174837

========================= Solving =========================
Time to solve HODLR form           :0.195398
Time taken to solve exactly        :0.0639205
Magnitude of Speed-Up              :0.32713
Forward Error(HODLR)               :4.66641e-16
Forward Error(Exact)               :5.74886e-16
Backward Error(HODLR)              :4.19436e-16
Backward Error(Exact)              :5.5721e-16

========================= Determinant Computation =========================
Time taken for HODLR form          :0.0341471
Calculated Log Determinant(HODLR)  :9165.94
Calculated Log Determinant(Exact)  :9165.94
Relative Error in computation      :1.02141e-14
```

As an aside, it seems that all uppercase variable names are being used in the `HODLR_Tree::plotTree` and one of these is shadowing the C `FILE` keyword. (I am curious: how portable is the direct `system` call?) With that said, the generated images are quite informative!

I am also doing my best not to veer into anything approaching a code review, but I would recommend not incorporating any `using` statements (e.g., `using std::cout;`) outside of a custom namespace.
