### 2019-01-20 48b00fa
Fixes few issues pointed out in review
- Now the example files use default argument values when the user doesn't provide them instead of segfaulting.
- Change of capital variable names in plotTree
- No longer is 'using std::cout' and 'using std::endl' used. However, using statements are still used for Mat, Vec, dtype and dtype_base
