### 2019-01-20 f80740e
Change to test file. Failure due to using cout, instead of std::cout
