### 2019-01-21 #1167 @sivaramambikasaran
@poulson Thanks. Our replies are inline.

> I am now noticing that running a standard build and then executing `./tutorial` leads to a segfault. It seems that one must manually specify the matrix height (`N`), leaf cutoff size (`M`), off-diagonal block rank (`dim`), and -log10(tolerance) as the first 4 arguments to `tutorial` to avoid a segfault. I would recommend using default versions of these parameters (or exiting with an error message) rather than segfaulting (and perhaps making note in the installation notes next to the cmake instructions).

This has been done. Unless all the four arguments are now passed, default arguments are used. Also, these arguments are relevant only for this specific example/tutorial file as has been mentioned [here](https://hodlrlib.readthedocs.io/en/latest/tutorial.html#running-the-program).

Further, the `dim` variable (i.e., the third input parameter) is **not** the off-diagonal rank. It is the dimension of the underlying data (Note that this is only needed for this tutorial file or in cases where the underlying matrix arises out of a kernel function and you need to sort the points in the appropriate dimension to get good low rank structure in the off-diagonal blocks). The ranks of the off-diagonal blocks are dictated by the  tolerance, i.e., the fourth argument.

>
> I am also wondering what the recommended example is, as the runtimes for HODLR seem to be significantly slower than dense linear algebra up to a matrix dimension of 4000. For example:
>
> ```
> ~/Source/JOSS/HODLR/build-full$ ./tutorial 4000 128 10 12

The reason for no speed up is that in the above case, the underlying dimension of the data is 10 and it turns out that the off-diagonal blocks are full rank. We suggest trying a lower dimension (say 1 or 2 for the current tutorial file). When the off-diagonal blocks are full-rank, our code will take a longer time than the usual LU, since our code tries to compress these full-rank off-diagonal blocks in the first place and also relies on a different factorization.

> As an aside, it seems that all uppercase variable names are being used in the `HODLR_Tree::plotTree` and one of these is shadowing the C `FILE` keyword. (I am curious: how portable is the direct `system` call?) With that said, the generated images are quite informative!
>
All variable names have been changed to lower-case in line with the rest of the code. The `system` command seems to work on the systems we have tested so far (Ubuntu 18.10, 16.04 and also on Mac Mojave 10.14.1).

> I am also doing my best not to veer into anything approaching a code review, but I would recommend not incorporating any `using` statements (e.g., `using std::cout;`) outside of a custom namespace in your header files.
>
This has now been fixed for cout and endl. We still rely on `using` for setting the datatype used in the code (i.e., float or double or complex32 or complex64).

> EDIT: I now realize that you are using a random matrix, so that the rank will be very high. I would recommend using your Gaussian random kernel example for your tutorial so that the user gets to see a speedup (the ultimate point of HODLR) as soon as possible.

As mentioned above, we use the Gaussian kernel on a random set of points. Just that the dimension was high, it resulted in full-rank off-diagonal blocks for these "small" matrices. We have used a random matrix as part of /test/test_HODLR.cpp to ensure that our solver is robust even for a random matrix with full-rank off-diagonal blocks.
