### 2019-01-21 #1167 @poulson
@sivaramambikasaran Unfortunately, that changeset broke the other examples, e.g., `examples/example.cpp` (namely, they were expecting `std::cout` to be imported into the global namespace). Along this line, is there a reason for forcing CMake to only build one executable for each configuration? I would highly recommend building all of the examples and tests with the default configuration; it would make it much easier for users (and myself) to explore the functionality and make sure everything builds.
