### 2019-01-21 9ae8645
Correction to files under examples/
