### 2019-01-22 a603a7c
Now all files under examples and test get built automatically
