### 2019-01-22 27d6a69
Fix to prevent crash on Travis
- When coverage build is performed, test subdirectory should not be build
- Update to documentation for verifying that the library is working as intended.
