### 2019-01-22 f2ce5a1
improve example
* Remove initialization with special seed of the PRNG. This was for debugging
  and shouldn't be included in the example anymore.
* Simplify kernel; in particular, compute the matrix elements inside the class.
* Add further comments and discussion.
