### 2019-01-22 #1167 @sivaramambikasaran
@poulson We have now fixed the other examples and the CMake builds all the examples and test files. Thanks.
