### 2019-01-22 99423de
Added printing of problem parameters to the example files
