### 2019-01-26 82c24f9
Now the low-rank matrix factorization part has been separated out and can be used independantly
