+++ src/Matrix_Factorizer.cpp
@@ -1,0 +1,357 @@
+#include "Matrix_Factorizer.hpp"
+ 
+void Matrix_Factorizer::maxAbsVector(const Vec& v, 
+                                        const std::set<int>& allowed_indices,
+                                        dtype& max, int& index
+                                       ) 
+{
+    std::set<int>::iterator it;
+    index = *allowed_indices.begin();
+    max   = v(index);
+
+    for(it = allowed_indices.begin(); it != allowed_indices.end(); it++) 
+    {
+        if(fabs(v(*it))>fabs(max)) 
+        {
+            index   =   *it;
+            max     =   v(index);
+        }
+    }
+}
+
+void Matrix_Factorizer::rookPiv(int n_row_start, int n_col_start, 
+                                   int n_rows, int n_cols, double tolerance, 
+                                   Mat& L, Mat& R, int& computed_rank
+                                  )
+{
+    // Indices which have been used:
+    std::vector<int> row_ind;      
+    std::vector<int> col_ind;
+
+    // Indices that are remaining:
+    std::set<int> remaining_row_ind;
+    std::set<int> remaining_col_ind;
+    
+    // Bases:
+    std::vector<Vec> u; 
+    std::vector<Vec> v;
+
+    for(int k = 0; k < n_rows; k++) 
+    {
+        remaining_row_ind.insert(k);
+    }
+    
+    for(int k = 0; k < n_cols; k++) 
+    {
+        remaining_col_ind.insert(k);
+    }
+
+    dtype max, gamma;
+
+    // Initialize the matrix norm and the the first row index
+    dtype_base matrix_norm = 0;
+    row_ind.push_back(0);
+    remaining_row_ind.erase(0);
+
+    // Stores the pivot entry of the considered row / col:
+    int pivot;
+
+    // This would get updated:
+    computed_rank = 0;
+    Vec row, col;
+    // These quantities in finding the stopping criteria:
+    dtype_base row_squared_norm, row_norm, col_squared_norm, col_norm;
+
+    // So these would be particularly useful for poorly conditioned matrices:
+    int max_tries = 10;
+    int count;
+
+    // Repeat till the desired tolerance is obtained
+    do 
+    {
+        // Generation of the row
+        // Row of the residuum and the pivot column
+        // By calling row_ind.back(), we are getting the last pushed number
+        row = this->A->getRow(n_row_start + row_ind.back(), n_col_start, n_cols);
+        
+        for(int i = 0; i < computed_rank; i++) 
+        {
+            row = row - u[i](row_ind.back()) * v[i];
+        }
+
+        this->maxAbsVector(row, remaining_col_ind, max, pivot);
+        count = 0;
+
+        // Alternating upon each call:
+        bool eval_at_end = false;
+        // Toggling randomness
+        bool use_randomization = true;
+
+        // This while loop is needed if in the middle of the algorithm the 
+        // row happens to be exactly the linear combination of the previous rows 
+        // upto some tolerance. i.e. prevents from ACA throwing false positives
+        while (fabs(max) < tolerance && 
+               count < max_tries   && 
+               remaining_col_ind.size() > 0 && 
+               remaining_row_ind.size() > 0
+              ) 
+        {
+            row_ind.pop_back();
+            int new_row_ind;
+
+            // When rank < 3, we will just choose entries from the ends of the matrix:
+            if(computed_rank < 3)
+            {
+                if(eval_at_end == true)
+                {
+                    new_row_ind = *remaining_row_ind.end();
+                }
+
+                else
+                {
+                    new_row_ind = *remaining_row_ind.begin();
+                }
+    
+                eval_at_end = !(eval_at_end);
+            }
+
+            // However, when we have rank >=3, we will choose the entries such that
[…]
