### 2019-01-28 ddfef44
Added SVD under factorization routines
