### 2019-01-28 061262d
WIP changes made for queen pivoting
