+++ src/Matrix_Factorizer.cpp
@@ -330,6 +330,150 @@ void Matrix_Factorizer::rookPiv(Mat& L, Mat& R, double tolerance,
     }
 }
 
+// void Matrix_Factorizer::queenPiv(Mat& L, Mat& R, double tolerance,
+//                                  int n_row_start, int n_col_start,
+//                                  int n_rows, int n_cols
+//                                 )
+// {
+//     // Indices which have been used:
+//     std::vector<int> row_ind;      
+//     std::vector<int> col_ind;
+
+//     // Indices that are remaining:
+//     std::set<int> remaining_row_ind;
+//     std::set<int> remaining_col_ind;
+    
+//     // Bases:
+//     std::vector<Vec> u; 
+//     std::vector<Vec> v;
+
+//     for(int k = 0; k < n_rows; k++) 
+//     {
+//         remaining_row_ind.insert(k);
+//     }
+    
+//     for(int k = 0; k < n_cols; k++) 
+//     {
+//         remaining_col_ind.insert(k);
+//     }
+
+//     dtype max, gamma;
+
+//     // Initialize the matrix norm and the the first row index
+//     dtype_base matrix_norm = 0;
+//     row_ind.push_back(0);
+//     remaining_row_ind.erase(0);
+
+//     // Stores the pivot entry of the considered row / col:
+//     int pivot;
+
+//     // This would get updated:
+//     int computed_rank = 0;
+//     Vec row, col;
+//     // These quantities in finding the stopping criteria:
+//     dtype_base row_squared_norm, row_norm, col_squared_norm, col_norm;
+
+//     // Repeat till the desired tolerance is obtained
+//     do 
+//     {
+//         diag1 = this->A->getDiag1(n_row_start + row_ind.back(), n_col_start + col_ind.back(), n_rows, n_cols);
+//         for(int i = 0; i < computed_rank; i++)
+//         {
+//             int row_idx, col_idx;
+//             #pragma omp parallel for 
+//             for (int j = 0; j < diag1.size(); j++)
+//             {
+//                 row_idx = n_row_start + j;
+//                 col_idx = n_col_start + j;
+
+//                 while(col_idx >= n_cols)
+//                     col_idx = col_idx - n_cols;
+//                 while(row_idx >= n_rows)
+//                     row_idx = row_idx - n_rows;
+//                 while(col_idx < 0)
+//                     col_idx = col_idx + n_cols;
+//                 while(row_idx < 0)
+//                     row_idx = row_idx + n_rows;
+
+//                 diag1(j) = diag1(j) - u[i](row_idx) * v[i](col_idx);
+//             }
+//         }
+
+//         if(n_rows > n_cols)
+//             this->maxAbsVector(diag, remaining_row_ind, max, pivot);
+//         else
+//             this->maxAbsVector(diag, remaining_col_ind, max, pivot);
+
+//         int new_row_ind, new_col_ind;
+//         new_col_ind = n_col_start + pivot;
+
+//         col_ind.push_back(new_col_ind);
+//         remaining_col_ind.erase(new_col_ind);
+//         // Normalizing constant
+//         gamma = dtype_base(1.0) / max;
+//         row_ind.push_back(n_row_start + pivot);
+//         remaining_row_ind.erase(n_row_start + pivot);
+
+//         // New vectors
+//         u.push_back(gamma * col);
+//         v.push_back(row);
+
+//         // New approximation of matrix norm
+//         row_squared_norm = row.squaredNorm();
+//         row_norm         = sqrt(row_squared_norm);
+
+//         col_squared_norm = col.squaredNorm();
+//         col_norm         = sqrt(col_squared_norm);
+
+//         // Updating the matrix norm:
[…]
