### 2019-01-29 a89b96d
WIP queen pivoting; Commiting version with a lot of printing to indicate what's going on
