+++ src/Matrix_Factorizer.cpp
@@ -330,149 +330,274 @@ void Matrix_Factorizer::rookPiv(Mat& L, Mat& R, double tolerance,
     }
 }
 
-// void Matrix_Factorizer::queenPiv(Mat& L, Mat& R, double tolerance,
-//                                  int n_row_start, int n_col_start,
-//                                  int n_rows, int n_cols
-//                                 )
-// {
-//     // Indices which have been used:
-//     std::vector<int> row_ind;      
-//     std::vector<int> col_ind;
+void Matrix_Factorizer::queenPiv(Mat& L, Mat& R, double tolerance,
+                                 int n_row_start, int n_col_start,
+                                 int n_rows, int n_cols
+                                )
+{
+    // Indices which have been used:
+    std::vector<int> row_ind;      
+    std::vector<int> col_ind;
 
-//     // Indices that are remaining:
-//     std::set<int> remaining_row_ind;
-//     std::set<int> remaining_col_ind;
+    // Indices that are remaining:
+    std::set<int> remaining_row_ind;
+    std::set<int> remaining_col_ind;
     
-//     // Bases:
-//     std::vector<Vec> u; 
-//     std::vector<Vec> v;
+    // Bases:
+    std::vector<Vec> u; 
+    std::vector<Vec> v;
 
-//     for(int k = 0; k < n_rows; k++) 
-//     {
-//         remaining_row_ind.insert(k);
-//     }
+    for(int k = 0; k < n_rows; k++) 
+    {
+        remaining_row_ind.insert(k);
+    }
     
-//     for(int k = 0; k < n_cols; k++) 
-//     {
-//         remaining_col_ind.insert(k);
-//     }
+    for(int k = 0; k < n_cols; k++) 
+    {
+        remaining_col_ind.insert(k);
+    }
 
-//     dtype max, gamma;
+    dtype max1, max2, max, gamma;
 
-//     // Initialize the matrix norm and the the first row index
-//     dtype_base matrix_norm = 0;
-//     row_ind.push_back(0);
-//     remaining_row_ind.erase(0);
+    // Initialize the matrix norm:
+    dtype_base matrix_norm = 0;
+    // Stores the pivot entry of the considered row / col:
+    int pivot1, pivot2, pivot;
 
-//     // Stores the pivot entry of the considered row / col:
-//     int pivot;
+    // This would get updated:
+    int computed_rank = 0;
+    Vec diag1, diag2, row, col;
+    // These quantities in finding the stopping criteria:
+    dtype_base row_squared_norm, row_norm, col_squared_norm, col_norm;
 
-//     // This would get updated:
-//     int computed_rank = 0;
-//     Vec row, col;
-//     // These quantities in finding the stopping criteria:
-//     dtype_base row_squared_norm, row_norm, col_squared_norm, col_norm;
+    // Repeat till the desired tolerance is obtained
+    do 
+    {
+        std::cout << "Iteration Number:" << computed_rank << std::endl << std::endl;
+        if(computed_rank == 0)
+        {
+            diag1 = this->A->getDiag1(n_row_start, n_col_start, n_rows, n_cols);
+        }
 
-//     // Repeat till the desired tolerance is obtained
-//     do 
-//     {
-//         diag1 = this->A->getDiag1(n_row_start + row_ind.back(), n_col_start + col_ind.back(), n_rows, n_cols);
-//         for(int i = 0; i < computed_rank; i++)
-//         {
-//             int row_idx, col_idx;
-//             #pragma omp parallel for 
-//             for (int j = 0; j < diag1.size(); j++)
-//             {
-//                 row_idx = n_row_start + j;
-//                 col_idx = n_col_start + j;
+        else
+        {
[…]
