+++ src/HODLR_Matrix.cpp
@@ -29,31 +29,21 @@ Vec HODLR_Matrix::getDiag1(const int n_row_start, const int n_col_start,
 {
     int N = std::max(n_rows, n_cols);
     Vec diag(N);
-    
+
     int row_ind, col_ind;
     #pragma omp parallel for
     for (int j = 0; j < N; ++j) 
     {   
         if(n_cols > n_rows)
         {
-            row_ind = n_row_start - n_col_start + j;
+            row_ind = this->mod(n_row_start - n_col_start + j, n_rows);
             col_ind = j;
-
-            while(row_ind >= n_rows)
-                row_ind = row_ind - n_rows;
-            while(row_ind < 0)
-                row_ind = row_ind + n_rows;
         }
 
         else
         {
             row_ind = j;
-            col_ind = n_col_start - n_row_start + j;
-    
-            while(col_ind >= n_cols)
-                col_ind = col_ind - n_cols;
-            while(col_ind < 0)
-                col_ind = col_ind + n_cols;
+            col_ind = this->mod(n_col_start - n_row_start + j, n_cols);
         }
         
         diag(j) = this->getMatrixEntry(row_ind, col_ind);
@@ -74,24 +64,14 @@ Vec HODLR_Matrix::getDiag2(const int n_row_start, const int n_col_start,
     {   
         if(n_cols > n_rows)
         {
-            row_ind = n_row_start + n_col_start - j;
+            row_ind = this->mod(n_row_start + n_col_start - j, n_rows);
             col_ind = j;
-
-            while(row_ind >= n_rows)
-                row_ind = row_ind - n_rows;
-            while(row_ind < 0)
-                row_ind = row_ind + n_rows;
         }
 
         else
         {
             row_ind = j;
-            col_ind = n_col_start + n_row_start - j;
-    
-            while(col_ind >= n_cols)
-                col_ind = col_ind - n_cols;
-            while(col_ind < 0)
-                col_ind = col_ind + n_cols;
+            col_ind = this->mod(n_col_start + n_row_start - j, n_cols);
         }
 
         diag(j) = this->getMatrixEntry(row_ind, col_ind);
