+++ examples/example_lowrank_factorization.cpp
@@ -15,8 +15,8 @@ public:
     // Constructor:
     Kernel(int N) : HODLR_Matrix(N) 
     {
-        Vec a = Vec::Random(N);
-        Vec b = Vec::Random(N);
+        Mat a = Mat::Random(N, 2);
+        Mat b = Mat::Random(N, 2);
 
         x = a * b.transpose();
     };
@@ -40,7 +40,7 @@ int main(int argc, char* argv[])
         std::cout << "All arguments weren't passed to executable!" << std::endl;
         std::cout << "Using Default Arguments:" << std::endl;
         // Size of the Matrix in consideration:
-        N          = 1000;
+        N          = 5;
         // Tolerance of problem
         tolerance  = pow(10, -12);
     }
@@ -59,12 +59,15 @@ int main(int argc, char* argv[])
 
     // Declaration of HODLR_Matrix object that abstracts data in Matrix:
     Kernel* K            = new Kernel(N);
-    Matrix_Factorizer* F = new Matrix_Factorizer(K);
+    Matrix_Factorizer* F = new Matrix_Factorizer(K, "queenPivoting");
+
+    Mat B = K->getMatrix(0, 0, N, N);
+    std::cout << "The input Matrix that has been provided is:" << std::endl;
+    std::cout << B << std::endl << std::endl << std::endl;
 
     Mat L, R;
     F->getFactorization(L, R, tolerance);
 
-    Mat B     = K->getMatrix(0, 0, N, N);
     Mat error = B - L * R.transpose();
     std::cout << "Accuracy of Factorization:" << error.maxCoeff() << std::endl << std::endl;
 
