### 2019-01-30 4b229f9
WIP changes made for matrix factorizer
