+++ src/Matrix_Factorizer.cpp
@@ -19,7 +19,7 @@ void Matrix_Factorizer::maxAbsVector(const Vec& v,
     }
 }
 
-void Matrix_Factorizer::rookPiv(Mat& L, Mat& R, double tolerance,
+void Matrix_Factorizer::rookPiv(Mat& L, Mat& R, double tolerance_or_rank,
                                 int n_row_start, int n_col_start,
                                 int n_rows, int n_cols
                                )
@@ -56,17 +56,25 @@ void Matrix_Factorizer::rookPiv(Mat& L, Mat& R, double tolerance,
     // Stores the pivot entry of the considered row / col:
     int pivot;
 
+    int target_rank = 0;
     // This would get updated:
     int computed_rank = 0;
     Vec row, col;
+
+    double tolerance = 0;
     // These quantities in finding the stopping criteria:
     dtype_base row_squared_norm, row_norm, col_squared_norm, col_norm;
+
+    if(tolerance_or_rank < 1)
+        tolerance = tolerance_or_rank;
+    else
+        target_rank = tolerance_or_rank;
 
     // So these would be particularly useful for poorly conditioned matrices:
     int max_tries = 10;
     int count;
 
-    // Repeat till the desired tolerance is obtained
+    // Repeat till the desired tolerance / rank is obtained
     do 
     {
         // Generation of the row
@@ -291,9 +299,12 @@ void Matrix_Factorizer::rookPiv(Mat& L, Mat& R, double tolerance,
         }
         
         computed_rank++;
-    } 
-    while(computed_rank * (n_rows + n_cols) * row_norm * col_norm > 
-          fabs(max) * tolerance * matrix_norm && 
+    }
+    while(((tolerance_or_rank < 1) ? 
+          computed_rank * (n_rows + n_cols) * row_norm * col_norm > 
+          fabs(max) * tolerance * matrix_norm 
+          : computed_rank < target_rank)
+          && 
           computed_rank < fmin(n_rows, n_cols)
          );
 
@@ -330,7 +341,7 @@ void Matrix_Factorizer::rookPiv(Mat& L, Mat& R, double tolerance,
     }
 }
 
-void Matrix_Factorizer::queenPiv(Mat& L, Mat& R, double tolerance,
+void Matrix_Factorizer::queenPiv(Mat& L, Mat& R, double tolerance_or_rank,
                                  int n_row_start, int n_col_start,
                                  int n_rows, int n_cols
                                 )
@@ -364,35 +375,28 @@ void Matrix_Factorizer::queenPiv(Mat& L, Mat& R, double tolerance,
     // Stores the pivot entry of the considered row / col:
     int pivot1, pivot2, pivot;
 
+    int target_rank = 0;
     // This would get updated:
     int computed_rank = 0;
     Vec diag1, diag2, row, col;
+
+    double tolerance = 0;
     // These quantities in finding the stopping criteria:
     dtype_base row_squared_norm, row_norm, col_squared_norm, col_norm;
 
-    // Repeat till the desired tolerance is obtained
+    if(tolerance_or_rank < 1)
+        tolerance = tolerance_or_rank;
+    else
+        target_rank = tolerance_or_rank;
+
+    // Repeat till the desired tolerance / rank is obtained
     do 
     {
-        std::cout << "Iteration Number:" << computed_rank << std::endl << std::endl;
         if(computed_rank == 0)
-        {
             diag1 = this->A->getDiag1(n_row_start, n_col_start, n_rows, n_cols);
-        }
-
         else
-        {
[…]
