### 2019-01-30 ac1cd90
WIP; Changes to look into geometric insight of ACA
