### 2019-01-30 #1167 @poulson
Thanks @sivaramambikasaran! It seems the only item left is clarifying what the recommended means of users contributing to the software is. To some degree, this is implicit in the GitHub model, but it would be good to have an explicit statement of the recommended process.
