### 2019-01-31 6ba00a9
Update CONTRIBUTING.md
