### 2019-02-01 #1167 @sivaramambikasaran
@poulson We have now added CONTRIBUTING.md file. Let us know if you would like us to add more information in that file.
