### 2019-02-02 #1167 @jedbrown
@sivaramambikasaran Can you add a DOI for the published paper? Would it make sense to include any citations to related work (software or otherwise) to compare capability, scope, and the like?

> `in these articles[@ambikasaran2013mathcal][@ambikasaran2014fast].`

should probably be

> `in @ambikasaran2013mathcal and @ambikasaran2014fast.`

You can remove the link to the repository; that information will be in template.  Also, I think your wild escapes are causing Pandoc to eat the title of the paper. Does normal math syntax not work?
