+++ docs/paper/paper.bib
@@ -6,6 +6,7 @@
   number={3},
   pages={477--501},
   year={2013},
+  doi={https://doi.org/10.1007/s10915-013-9714-z},
   publisher={Springer}
 }
 
@@ -13,5 +14,6 @@
   title={Fast symmetric factorization of hierarchical matrices with applications},
   author={Ambikasaran, Sivaram and O'Neil, Michael and Singh, Karan Raj},
   journal={arXiv preprint arXiv:1405.0223},
+  doi={arXiv:1405.0223},
   year={2014}
 }
