+++ docs/paper/paper.bib
@@ -12,8 +12,16 @@
 
 @article{ambikasaran2014fast,
   title={Fast symmetric factorization of hierarchical matrices with applications},
-  author={Ambikasaran, Sivaram and O'Neil, Michael and Singh, Karan Raj},
-  journal={arXiv preprint arXiv:1405.0223},
-  doi={arXiv:1405.0223},
-  year={2014}
+  author={{Ambikasaran}, Sivaram and {O'Neil}, Michael and {Singh}, Karan Raj},
+  journal={arXiv e-prints},
+  keywords={Mathematics - Numerical Analysis, Computer Science - Numerical Analysis, Physics - Fluid Dynamics, Statistics - Computation},
+  year={2014},
+  month={May},
+  eid={arXiv:1405.0223},
+  pages={arXiv:1405.0223},
+  archivePrefix={arXiv},
+  eprint={1405.0223},
+  primaryClass={math.NA},
+  adsurl={https://ui.adsabs.harvard.edu/\#abs/2014arXiv1405.0223A},
+  adsnote={Provided by the SAO/NASA Astrophysics Data System}
 }
