### 2019-02-04 #1167 @sivaramambikasaran
@jedbrown Our replies are inline.

> @sivaramambikasaran Can you add a DOI for the published paper?

Done.

> Would it make sense to include any citations to related work (software or otherwise) to compare capability, scope, and the like?

While our library has some similarity with other libraries such as STRUMPACK and H2Lib, each works on a different structure and comparisons among them would be difficult to evaluate.

> > `in these articles[@ambikasaran2013mathcal][@ambikasaran2014fast].`
>
> should probably be
>
> > `in @ambikasaran2013mathcal and @ambikasaran2014fast.`

Done.

> You can remove the link to the repository; that information will be in template.

Done.

> Also, I think your wild escapes are causing Pandoc to eat the title of the paper. Does normal math syntax not work?

We tried compiling using pandoc from markdown to LaTeX and it seems to work fine. The article generated by whedon also shows the title. So we don't quite understand where/how the title of the paper is eaten.
