+++ docs/paper/paper.bib
@@ -25,3 +25,119 @@
   adsurl={https://ui.adsabs.harvard.edu/\#abs/2014arXiv1405.0223A},
   adsnote={Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@article {hmatrix1999,
+author = {Hackbusch, Wolfgang},
+title = {{A sparse matrix arithmetic based on $\mathcal{H}$-matrices. I. Introduction to $\mathcal{H}$ matrices}},
+doi = {10.1007/s006070050015},
+journal = {computing},
+pages = {89--108},
+year = {1999},
+volume = {62},
+number = {2},
+issn = {0010-485x},
+institution = {{Max Planck Institute for Mathematics in the Sciences}},
+}
+
+@article {hmatrix2000,
+author = {Hackbusch, Wolfgang and Khoromsky, Boris N.},
+title = {{sparse $\mathcal{H}$ - matrix arithmetic. Part II. Application to multi-dimensional
+problems}},
+doi = {10.1007/s006070050002},
+journal = {computing},
+pages = {21--47},
+year = {2000},
+volume = {64},
+number = {1},
+issn = {0010-485x},
+institution = {{Max Planck Institute for Mathematics in the Sciences}},
+MISPreprint = {{MIS-Preprint 22/1999}},
+url = {http://www.mis.mpg.de/de/publications/preprints/1999/prepr1999-22.html}, 
+}
+
+@article {h2matrix2002,
+author = {Hackbusch, Wolfgang and Borm, Steffen},
+title = {{Data sparse approximation by adaptive $\mathcal{H}^2$ - matrices}},
+doi = {10.1007/s00607-002-1450-4},
+journal = {computing},
+pages = {1--35},
+year = {2002},
+volume = {69},
+number = {1},
+issn = {0010-485x},
+institution = {{Max Planck Institute for Mathematics in the Sciences}},
+MISPreprint = {{MIS-Preprint 86/2001}},
+url = {http://www.mis.mpg.de/de/publications/preprints/2001/prepr2001-86.html}, 
+}
+
+@article{chandrasekaran2006fast,
+  title={A fast ULV decomposition solver for hierarchically semiseparable representations},
+  author={Chandrasekaran, Shiv and Gu, Ming and Pals, Timothy},
+  journal={SIAM Journal on Matrix Analysis and Applications},
+  volume={28},
+  number={3},
+  pages={603--622},
+  year={2006},
+  publisher={SIAM}
+}
+
+@phdthesis{ambikasaran2013fast,
+  title={Fast algorithms for dense numerical linear algebra and applications},
+  author={Ambikasaran, Sivaram},
+  year={2013},
+  school={Stanford University Stanford}
+}
+
+@article{kressner2017fast,
+  title={Fast computation of spectral projectors of banded matrices},
+  author={Kressner, Daniel and Susnjara, Ana},
+  journal={SIAM Journal on Matrix Analysis and Applications},
+  volume={38},
+  number={3},
+  pages={984--1009},
+  year={2017},
+  publisher={SIAM}
+}
+
+@article{Hartmann2018,
+	doi = {10.1088/1402-4896/aae34e},
+	url = {https://doi.org/10.1088%2F1402-4896%2Faae34e},
+	year = 2018,
+	month = {oct},
+	publisher = {{IOP} Publishing},
+	volume = {93},
+	number = {11},
+	pages = {114003},
+	author = {Michael Hartmann and Gert-Ludwig Ingold and Paulo A Maia Neto},
+	title = {Advancing numerics for the Casimir effect to experimentally relevant aspect ratios},
+	journal = {Physica Scripta}
+}
+
+@article{ambikasaran2016fast,
+  title={Fast direct methods for Gaussian processes},
[…]
