+++ docs/paper/paper.bib
@@ -71,7 +71,8 @@ url = {http://www.mis.mpg.de/de/publications/preprints/2001/prepr2001-86.html},
 }
 
 @article{chandrasekaran2006fast,
-  title={A fast ULV decomposition solver for hierarchically semiseparable representations},
+  title={{A fast $ULV$ decomposition solver for hierarchically semiseparable representations}},
+  doi={10.1137/S0895479803436652},
   author={Chandrasekaran, Shiv and Gu, Ming and Pals, Timothy},
   journal={SIAM Journal on Matrix Analysis and Applications},
   volume={28},
@@ -90,6 +91,7 @@ url = {http://www.mis.mpg.de/de/publications/preprints/2001/prepr2001-86.html},
 
 @article{kressner2017fast,
   title={Fast computation of spectral projectors of banded matrices},
+  doi={10.1137/16M1087278},
   author={Kressner, Daniel and Susnjara, Ana},
   journal={SIAM Journal on Matrix Analysis and Applications},
   volume={38},
@@ -115,6 +117,7 @@ url = {http://www.mis.mpg.de/de/publications/preprints/2001/prepr2001-86.html},
 
 @article{ambikasaran2016fast,
   title={Fast direct methods for Gaussian processes},
+  doi={10.1109/TPAMI.2015.2448083},
   author={Ambikasaran, Sivaram and Foreman-Mackey, Daniel and Greengard, Leslie and Hogg, David W and O’Neil, Michael},
   journal={IEEE transactions on pattern analysis and machine intelligence},
   volume={38},
@@ -126,6 +129,7 @@ url = {http://www.mis.mpg.de/de/publications/preprints/2001/prepr2001-86.html},
 
 @article{ghysels2016efficient,
   title={An efficient multicore implementation of a novel HSS-structured multifrontal solver using randomized sampling},
+  doi={10.1137/15M1010117},
   author={Ghysels, Pieter and Li, Xiaoye S and Rouet, Fran{\c{c}}ois-Henry and Williams, Samuel and Napov, Artem},
   journal={SIAM Journal on Scientific Computing},
   volume={38},
@@ -138,6 +142,6 @@ url = {http://www.mis.mpg.de/de/publications/preprints/2001/prepr2001-86.html},
 @misc{rouet_2015, 
       title={A comparison of different low-rank approximation techniques},
       url={https://personalpages.manchester.ac.uk/staff/theo.mary/doc/LA15.pdf},
-      author={Rouet, François-Henry},
+      author={Rouet, Francois Henry},
       year={2015}
      }
