+++ docs/paper/paper.md
@@ -25,15 +25,15 @@ bibliography: paper.bib
 
 # Summary
 
-``HODLRlib`` is a flexible library for working with matrices that have a Hierarchical Off-Diagonal Low-Rank (HODLR) [@ambikasaran2013mathcal] structure. The current version performs matrix operations like matrix-vector products, solving linear systems, Cholesky-like symmetric factorization and determinant computation in almost linear complexity. 
+``HODLRlib`` is a flexible library for working with matrices that have a Hierarchical Off-Diagonal Low-Rank (HODLR) (@ambikasaran2013mathcal) structure. The current version performs matrix operations like matrix-vector products, solving linear systems, Cholesky-like symmetric factorization and determinant computation in almost linear complexity. 
 
-Significant speedups are obtained with ``HODLRlib``, when dealing with matrices that have an underlying data sparsity (rank-structured). Most rank-structured matrices fall in the class of Hierarchical matrices: some examples are $\mathcal{H}$[@hmatrix1999, @hmatrix2000], $\mathcal{H}^2$ [@h2matrix2002], HSS [@chandrasekaran2006fast], HODLR. These formats differ from each other in their matrix partitioning, low-rank sub-blocks and nested basis behaviour. A thorough comparison of these different formats has been provided in [@ambikasaran2013fast].
+Significant speedups are obtained with ``HODLRlib``, when dealing with matrices that have an underlying data sparsity (rank-structured). Most rank-structured matrices fall in the class of Hierarchical matrices: some examples are $\mathcal{H}$ (@hmatrix1999, @hmatrix2000), $\mathcal{H}^2$ (@h2matrix2002), HSS (@chandrasekaran2006fast), HODLR. These formats differ from each other in their matrix partitioning, low-rank sub-blocks and nested basis behaviour. A thorough comparison of these different formats has been provided in [@ambikasaran2013fast].
 
 Among these hierarchical matrices, HODLR matrices are applicable for a wide range of problems of practical interest [@kressner2017fast, @Hartmann2018, @ambikasaran2016fast]. In addition, they possess the simplest data-sparse representation, which inturn allows for easier implementation and obtaining great speed-ups.
 
-``HODLRlib`` strives to provide a functionality for ``HODLR`` matrices similar to how libraries such as STRUMPACK [@ghysels2016efficient] and H2lib [@h2matrix2002] have provided for the HSS and $\mathcal{H}^2$ matrix formats respectively. While it is difficult to determine which matrix format works best, studies @rouet_2015 in the past have revealed that it isn't a case of "one size fits all". Rather, it reveals that the optimum hierarchical matrix format is dependent on the problem and the size considered.
+``HODLRlib`` strives to provide a functionality for ``HODLR`` matrices similar to how libraries such as STRUMPACK (@ghysels2016efficient) and H2lib (@h2matrix2002) have provided for the HSS and $\mathcal{H}^2$ matrix formats respectively. Some unique features of our library are Cholesky-like symmetric factorization and determinant computation. While it is difficult to determine which matrix format works best, studies @rouet_2015 in the past have revealed that it isn't a case of "one size fits all". Rather, it reveals that the optimum hierarchical matrix format is dependent on the problem and the size considered.
 
[…]
