+++ docs/paper/paper.bib
@@ -86,7 +86,7 @@ url = {http://www.mis.mpg.de/de/publications/preprints/2001/prepr2001-86.html},
   title={Fast algorithms for dense numerical linear algebra and applications},
   author={Ambikasaran, Sivaram},
   year={2013},
-  school={Stanford University Stanford}
+  school={Stanford University}
 }
 
 @article{kressner2017fast,
