### 2019-02-05 e0c3788
Update paper.md
Added brackets for rouet_2015
