+++ docs/paper/paper.md
@@ -31,7 +31,7 @@ Significant speedups are obtained with ``HODLRlib``, when dealing with matrices 
 
 Among these hierarchical matrices, HODLR matrices are applicable for a wide range of problems of practical interest [@kressner2017fast, @Hartmann2018, @ambikasaran2016fast]. In addition, they possess the simplest data-sparse representation, which inturn allows for easier implementation and obtaining great speed-ups.
 
-``HODLRlib`` strives to provide a functionality for ``HODLR`` matrices similar to how libraries such as STRUMPACK (@ghysels2016efficient) and H2lib (@h2matrix2002) have provided for the HSS and $\mathcal{H}^2$ matrix formats respectively. Some unique features of our library are Cholesky-like symmetric factorization and determinant computation. While it is difficult to determine which matrix format works best, studies @rouet_2015 in the past have revealed that it isn't a case of "one size fits all". Rather, it reveals that the optimum hierarchical matrix format is dependent on the problem and the size considered.
+``HODLRlib`` strives to provide a functionality for ``HODLR`` matrices similar to how libraries such as STRUMPACK (@ghysels2016efficient) and H2lib (@h2matrix2002) have provided for the HSS and $\mathcal{H}^2$ matrix formats respectively. Some unique features of our library are Cholesky-like symmetric factorization and determinant computation. While it is difficult to determine which matrix format works best, studies (@rouet_2015) in the past have revealed that it isn't a case of "one size fits all". Rather, it reveals that the optimum hierarchical matrix format is dependent on the problem and the size considered.
 
 ``HODLRlib`` is an optimized implementation of the ideas illustrated in [@ambikasaran2013mathcal, @ambikasaran2016fast, @ambikasaran2014fast]. The goal of ``HODLRlib`` is to serve as a high-performance, easy to use reference implementation for working with any HODLR matrix. We believe that apart from the features provided by the library, the simplicity and the extendability of the implementation is our USP. Our interfaces are kept simple with a minimum number of dependencies, with well-documented code to ensure that a potential user/developer can hit the ground running as soon as possible.
 
