### 2019-02-05 #1167 @sivaramambikasaran
@jedbrown We are done now (hopefully). One little thing is I do not know how to get doi of a thesis. Also, we have cited a SIAM LA presentation. Hopefully this is fine as well.
