### 2019-02-06 910992a
Minor changes to formatting; codacy fixes
