+++ CONTRIBUTING.md
@@ -4,10 +4,10 @@
 
 We both value and encourage the community to provide feedback about using HODLRlib and the issues they are facing. Issues can span a variety of topics including:
 
-  - Feature requests
-  - Bug reports
-  - Build Issues
-  - Performance Issues
+-  Feature requests
+-  Bug reports
+-  Build Issues
+-  Performance Issues
 
 Please use the Github [issue tracker](https://github.com/sivaramambikasaran/HODLR/issues) to file your issues
 
@@ -17,8 +17,8 @@ If you want to contribute code, we suggest you that use fork the repository, mak
 
 Key areas of code contributions include:
 
-  - New features
-  - Bug fixes
-  - Performance improvements
-  - We highly value new examples demonstrating usage of our code
-  - We are keen on creating wrappers in other languages to HODLRlib (at the moment, our interest is in wrappers for Python and Julia). If you are experienced in any of these languages, please do contribute!
+-  New features
+-  Bug fixes
+-  Performance improvements
+-  We highly value new examples demonstrating usage of our code
+-  We are keen on creating wrappers in other languages to HODLRlib (at the moment, our interest is in wrappers for Python and Julia). If you are experienced in any of these languages, please do contribute!
