### 2019-02-06 4fe484d
Small formatting changes
