### 2019-02-06 fb5d46f
Trying out different formatting for contributing.md
