### 2019-02-06 #1167 @jedbrown
There's a bib formatting issue, described here. https://github.com/sivaramambikasaran/HODLR/commit/1bef70dbf42d1748ab2471463eb46e2b14cadd7d

"Hierachical" is misspelled.

Looks good otherwise; thanks for this further explanation.
