### 2019-02-06 dfddf56
Made fixes to the paper
