### 2019-02-06 #1167 @sivaramambikasaran
@jedbrown

> There's a bib formatting issue, described here. [sivaramambikasaran/HODLR@1bef70d](https://github.com/sivaramambikasaran/HODLR/commit/1bef70dbf42d1748ab2471463eb46e2b14cadd7d)

Done

>
> "Hierachical" is misspelled.

Fixed
>
> Looks good otherwise; thanks for this further explanation.

Hopefully, she can fly now.
