### 2019-02-08 #1167 @jedbrown
Please consider this PR, then tag a release ([annotated tag](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) preferred) and archive on Zenodo or similar.

https://github.com/sivaramambikasaran/HODLR/pull/37
