+++ docs/paper/paper.md
@@ -33,7 +33,7 @@ Among these hierarchical matrices, HODLR matrices are applicable for a wide rang
 
 ``HODLRlib`` strives to provide functionality for ``HODLR`` matrices similar to that which libraries such as STRUMPACK [@ghysels2016efficient] and H2lib [@h2matrix2002] provide for the HSS and $\mathcal{H}^2$ matrix formats respectively. Some unique features of our library are Cholesky-like symmetric factorization and determinant computation. While it is difficult to determine which matrix format works best, studies [@rouet_2015] in the past have revealed that it isn't a case of "one size fits all". Rather, it reveals that the optimum hierarchical matrix format is dependent on the problem and the size considered.
 
-``HODLRlib`` is an optimized implementation of the ideas illustrated in @ambikasaran2013mathcal, @ambikasaran2016fast and @ambikasaran2014fast. The goal of ``HODLRlib`` is to serve as a high-performance, easy to use reference implementation for working with any HODLR matrix. We believe that apart from the features provided by the library, the simplicity and the extendability of the implementation is our USP. Our interfaces are kept simple with a minimum number of dependencies, with well-documented code to ensure that a potential user/developer can hit the ground running as soon as possible.
+``HODLRlib`` is an optimized implementation of the ideas illustrated in @ambikasaran2013mathcal, @ambikasaran2016fast and @ambikasaran2014fast. The goal of ``HODLRlib`` is to serve as a high-performance, easy to use reference implementation for working with any HODLR matrix. We believe that apart from the features provided by the library, the simplicity and the extendability of the implementation distinguishes ours from other libraries. Our interfaces are kept simple with a minimum number of dependencies, with well-documented code to ensure that a potential user/developer can hit the ground running as soon as possible.
 
 Our code makes use of shared-memory parallelism through OpenMP. The solver is fairly generic and can handle matrices not necessarily arising out of kernel functions. Further, the solver has been optimized and the running time of the solver is now massively (a few orders of magnitude) faster than the running times reported in the original articles.
 
