### 2019-02-11 #1167 @sivaramambikasaran
@jedbrown

> Please consider this PR,

Done

> then tag a release ([annotated tag](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) preferred)

Done

> and archive on Zenodo or similar.

Done

See here: https://zenodo.org/record/2561575#.XGETt88zYW8

DOI: 10.5281/zenodo.2561575
