### 2019-02-12 #1167 @labarba
👋 @sivaramambikasaran : Can you update the tile of the Zenodo entry so that it matches the title of the JOSS paper?
